extract the two steps of group creation into their own pages
