+++ ui/src/routes/new-group/+page.svelte
@@ -1,180 +1,35 @@
 <script lang="ts">
-	import '@awesome.me/webawesome/dist/components/icon/icon.js';
-	import { m, members } from '$lib/paraglide/messages.js';
-
 	import { getContext } from 'svelte';
 	import { goto } from '$app/navigation';
-	import type { ContactsStore, ChatsStore, PublicKey } from 'dash-chat-stores';
-	import { wrapPathInSvg } from '$lib/utils/icon';
-	import { mdiAccountMultiplePlus } from '@mdi/js';
+	import type { ChatsStore, PublicKey } from 'dash-chat-stores';
+	import MembersStep from './MembersStep.svelte';
+	import GroupInfoStep from './GroupInfoStep.svelte';
 
-	import { useReactivePromise } from '$lib/stores/use-signal';
-	import ProfileAvatar from '$lib/components/profiles/ProfileAvatar.svelte';
-	import SelectAvatar from '$lib/components/profiles/SelectAvatar.svelte';
-	import {
-		Page,
-		Navbar,
-		NavbarBackLink,
-		Button,
-		Card,
-		Link,
-		List,
-		ListItem,
-		Checkbox,
-		ListInput,
-		BlockTitle,
-		Preloader,
-		useTheme,
-	} from 'konsta/svelte';
-	import { mdiArrowNext } from '$lib/utils/icon';
-	import { isIos } from '$lib/utils/environment';
-	import { isWideScreen } from '$lib/stores/screen.svelte';
-
-	const contactsStore: ContactsStore = getContext('contacts-store');
 	const chatsStore: ChatsStore = getContext('chats-store');
-
-	const contacts = useReactivePromise(contactsStore.profilesForAllContacts);
 
 	let currentPage: 'members' | 'group-info' = $state('members');
 	let selectedContacts = $state<PublicKey[]>([]);
 	let groupName = $state('');
 	let groupAvatar = $state<string | undefined>(undefined);
-	const theme = $derived(useTheme());
 
 	async function createGroupChat() {
-		const contacts = Array.from(selectedContacts);
-		const groupStore = await chatsStore.createGroup(contacts);
+		const groupStore = await chatsStore.createGroup(
+			Array.from(selectedContacts),
+		);
 		goto(`/group-chat/${groupStore.chatId}`);
 	}
 </script>
 
 {#if currentPage === 'members'}
-	<Page>
-		<Navbar title={m.newGroup()} titleClass="opacity1" transparent={true}>
-			{#snippet left()}
-				<NavbarBackLink
-					onClick={() => window.history.back()}
-					data-testid="new-group-back"
-				/>
-			{/snippet}
-
-			{#snippet right()}
-				{#if isIos}
-					<Link
-						onClick={() => (currentPage = 'group-info')}
-						data-testid="new-group-next-link"
-					>
-						{selectedContacts.length === 0 ? m.skip() : m.next()}
-					</Link>
-				{/if}
-			{/snippet}
-		</Navbar>
-
-		<div class="column" style="flex: 1">
-			<div class="center-in-desktop">
-				<BlockTitle>{m.contacts()}</BlockTitle>
-
-				<List strongIos inset={isWideScreen.value || theme === 'ios'}>
-					{#await $contacts}
-						<div
-							class="column"
-							style="flex: 1; align-items: center; justify-content: center"
-						>
-							<Preloader />
-						</div>
-					{:then contacts}
-						{#each contacts as [publicKey, profile]}
-							<ListItem label title={profile.name}>
-								{#snippet media()}
-									<ProfileAvatar chatActorId={publicKey}></ProfileAvatar>
-								{/snippet}
-
-								{#snippet after()}
-									<Checkbox
-										checked={selectedContacts.includes(publicKey)}
-										onChange={e => {
-											const target = e.target as HTMLInputElement;
-											if (target.checked) {
-												selectedContacts = [...selectedContacts, publicKey];
-											} else {
-												selectedContacts = selectedContacts.filter(
-													c => c !== publicKey,
-												);
-											}
-										}}
-									/>
-								{/snippet}
-							</ListItem>
-						{:else}
-							<ListItem title={m.noContactsYet()} />
-						{/each}
-					{/await}
-				</List>
-			</div>
-		</div>
-
-		{#if !isIos}
-			<Button
-				onClick={() => (currentPage = 'group-info')}
-				data-testid="new-group-next-btn"
-				class="fixed-action-btn"
-				rounded
-			>
-				{selectedContacts.length === 0 ? m.skip() : m.next()}
-			</Button>
-		{/if}
-	</Page>
-{:else}
-	<Page>
-		<Navbar title={m.groupName()} titleClass="opacity1" transparent={true}>
-			{#snippet left()}
-				<NavbarBackLink
-					onClick={() => (currentPage = 'members')}
-					data-testid="new-group-info-back"
-				/>
-			{/snippet}
-
-			{#snippet right()}
-				{#if isIos}
-					<Link onClick={createGroupChat} data-testid="new-group-create-link">
-						{m.create()}
-					</Link>
-				{/if}
-			{/snippet}
-		</Navbar>
-
-		<div class="column" style="flex: 1">
-			<div class="center-in-desktop m-1">
-				<List
-					inset={isWideScreen.value || theme === 'ios'}
-					strongIos
-					nested={theme !== 'ios'}
-				>
-					<ListInput
-						type="text"
-						bind:value={groupName}
-						data-testid="new-group-name-input"
-						outline
-						class="plain"
-						placeholder={m.name()}
-					>
-						{#snippet media()}
-							<SelectAvatar bind:value={groupAvatar}></SelectAvatar>
-						{/snippet}
-					</ListInput>
-				</List>
-			</div>
-		</div>
-
-		{#if !isIos}
-			<Button
-				onClick={createGroupChat}
-				data-testid="new-group-create-btn"
-				class="fixed-action-btn"
-				rounded
-			>
-				{m.create()}
-			</Button>
-		{/if}
-	</Page>
+	<MembersStep
+		bind:selectedContacts
+		onNext={() => (currentPage = 'group-info')}
+	/>
+{:else if currentPage === 'group-info'}
+	<GroupInfoStep
+		bind:groupName
+		bind:groupAvatar
+		onBack={() => (currentPage = 'members')}
+		onCreate={createGroupChat}
+	/>
 {/if}
